ขาดlogin+noti+ใชขอมลเจคกลอง

+++ Models/EICC_Control.cs
@@ -8,9 +8,9 @@ namespace Api.Models
         public int ControlID { get; set; }
         public int EmpID { get; set; }
         public int WeekID { get; set; }
-        public float? TotalHours { get; set; }
+        public double? TotalHours { get; set; }
         public int? DaysWorked { get; set; }
-        public float? TotalOT { get; set; } // ✅ เพิ่มตรงนี้!!
+        public double? TotalOT { get; set; } // ✅ เพิ่มตรงนี้!!
         public string Status { get; set; }
     }
 }
